Not yet connected
I still have to establish a connection between the two databases but the bones of a file structure are here as well as the supbase node module.

The current goal is to establish a connection and then use the database to display the song library at the start of the application.

Only then will frontend work properly begin